Struct holochain_dpki::keypair::SigningKeyPair

pub struct SigningKeyPair {
    pub public: Base32,
    pub private: SecBuf,
}

KeyPair used for signing data

Fields
public: Base32private: SecBuf
Methods

impl SigningKeyPair

pub fn new(public: Base32, private: SecBuf) -> Self

Standard Constructor

pub fn new_with_raw_key(pub_key_sec: &mut SecBuf, private: SecBuf) -> Self

Construct with a public key not already HCID encoded

pub fn sign(&mut self, data: &mut SecBuf) -> HcResult<SecBuf>

sign some arbitrary data with the signing private key @param {SecBuf} data - the data to sign @return {SecBuf} signature - Empty SecBuf to be filled with the signature

pub fn verify(&mut self, data: &mut SecBuf, signature: &mut SecBuf) -> bool

verify data that was signed with our private signing key @param {SecBuf} data @param {SecBuf} signature @return true if verification succeeded